requirements.txt
streamlit==1.38.0
openai==1.44.1
PyPDF2==3.0.1
python-docx==0.8.11
pdf2image==1.16.3
pillow>=9.0.0  # Ensure you're using a recent version of Pillow
httpx>=0.23.0 # Added for explicit httpx version